feat: eth_bridge with tests in dedicated folder
IMI-2282
